authorization implemented

+++ src/api/notes/handler.js
@@ -16,8 +16,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = 'untitled', body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const noteId = await this._service.addNote({ title, body, tags, owner: credentialId});
 
       const response = h.response({
         status: 'success',
@@ -40,15 +41,16 @@ class NotesHandler {
 
       const response = h.response({
         status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        message: 'maaf, terjadi kegagalan pada server kami.',
       });
       response.code(500);
       console.error(error);
       return response;
     }
   }
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -60,7 +62,11 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id:credentialId } =request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
+
       return {
         status: 'success',
         data: {
@@ -79,7 +85,7 @@ class NotesHandler {
 
       const response = h.response({
         status: 'error',
-        message: 'Maaf, terjadi kesalahan pada server kami.',
+        message: 'maaf, terjadi kesalahan pada server kami.',
       });
       response.code(500);
       console.error(error);
@@ -91,12 +97,14 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.editNoteById(id, request.payload);
 
       return {
         status: 'success',
-        message: 'Catatan berhasil diperbarui',
+        message: 'catatan berhasil diperbarui',
       };
     } catch (error) {
       if (error instanceof ClientError) {
@@ -109,7 +117,7 @@ class NotesHandler {
       }
       const response = h.response({
         status: 'error',
-        message: 'Maaf, terjadi kesalahan pada server kami',
+        message: 'maaf, terjadi kesalahan pada server kami',
       });
       response.code(500);
       console.error(error);
@@ -120,11 +128,14 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id:credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
 
       return {
         status: 'success',
-        message: 'Catatan berhasil dihapus',
+        message: 'catatan berhasil dihapus',
       };
     } catch (error) {
       if (error instanceof ClientError) {
@@ -137,7 +148,7 @@ class NotesHandler {
       }
       const response = h.response({
         status: 'error',
-        message: 'Maaf, terjadi kesalahan pada server kami',
+        message: 'maaf, terjadi kesalahan pada server kami',
       });
       response.code(500);
       console.error(error);
